// project-Intership/src/app/shared/model/harvest-forest-filter.ts
export class HarvestForestFilterParam {
    fromDateNumber: number;
    toDateNumber: number;
    status: string;
    createFromForest: boolean;
}